import checkstyle xml file and update pom, rename parquet module(use lower case)
