+++ oap-ape/ape-java/ape-common/src/main/java/com/intel/ape/parquet/ApeContainsFilter.java
@@ -15,11 +15,10 @@
  * limitations under the License.
  */
 
-package com.intel.ape.Parquet;
+package com.intel.ape.parquet;
 
 import org.apache.parquet.filter2.predicate.Statistics;
 import org.apache.parquet.io.api.Binary;
-
 
 public class ApeContainsFilter extends ApeLikeFilter {
   public ApeContainsFilter(String value) {
