+++ oap-ape/ape-java/ape-common/src/main/java/com/intel/ape/parquet/ApeEndWithFilter.java
@@ -21,9 +21,8 @@ import org.apache.parquet.filter2.predicate.Statistics;
 import org.apache.parquet.io.api.Binary;
 import org.apache.parquet.schema.PrimitiveComparator;
 
-
-public class ApeStartWithFilter extends ApeLikeFilter {
-  public ApeStartWithFilter(String value) {
+public class ApeEndWithFilter extends ApeLikeFilter {
+  public ApeEndWithFilter(String value) {
     super(value);
   }
 
@@ -34,20 +33,26 @@ public class ApeStartWithFilter extends ApeLikeFilter {
 
   @Override
   public boolean canDrop(Statistics<Binary> statistics) {
-    PrimitiveComparator<Binary> comparator = PrimitiveComparator.UNSIGNED_LEXICOGRAPHICAL_BINARY_COMPARATOR;
+    PrimitiveComparator<Binary> comparator =
+            PrimitiveComparator.UNSIGNED_LEXICOGRAPHICAL_BINARY_COMPARATOR;
     Binary max = statistics.getMax();
     Binary min = statistics.getMin();
-    return comparator.compare(max.slice(0, Math.min(size, max.length())), strToBinary) < 0 ||
-            comparator.compare(min.slice(0, Math.min(size, min.length())), strToBinary) > 0;
+    int lenInMax = Math.min(size, max.length());
+    int lenInMin = Math.min(size, min.length());
+    return comparator.compare(max.slice(max.length() - lenInMax, lenInMax), strToBinary) < 0 ||
+            comparator.compare(min.slice(min.length() - lenInMin, lenInMin), strToBinary) > 0;
   }
 
   @Override
   public boolean inverseCanDrop(Statistics<Binary> statistics) {
-    PrimitiveComparator<Binary> comparator = PrimitiveComparator.UNSIGNED_LEXICOGRAPHICAL_BINARY_COMPARATOR;
+    PrimitiveComparator<Binary> comparator =
+            PrimitiveComparator.UNSIGNED_LEXICOGRAPHICAL_BINARY_COMPARATOR;
     Binary max = statistics.getMax();
     Binary min = statistics.getMin();
-    return comparator.compare(max.slice(0, Math.min(size, max.length())), strToBinary) == 0 &&
-            comparator.compare(min.slice(0, Math.min(size, min.length())), strToBinary) == 0;
+    int lenInMax = Math.min(size, max.length());
+    int lenInMin = Math.min(size, min.length());
+    return comparator.compare(max.slice(max.length() - lenInMax, lenInMax), strToBinary) == 0 &&
+            comparator.compare(min.slice(min.length() - lenInMin, lenInMin), strToBinary) == 0;
   }
 
 }
