+++ oap-ape/ape-java/pom.xml
@@ -24,6 +24,8 @@
     <slf4j.version>1.7.16</slf4j.version>
     <hadoop.version>2.4.1</hadoop.version>
     <ape.version>1.1.0-SNAPSHOT</ape.version>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
   </properties>
 
   <build>
